extract out useFetchData
