+++ src/containers/allelePage/useVariantTranscripts.js
@@ -1,41 +1,4 @@
-import { useState, useCallback } from 'react';
-import { buildTableQueryString } from '../../lib/utils';
-
-function useFetchData(url) {
-  const [data, setData] = useState({
-    data: [],
-    total: 0,
-  });
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(null);
-
-  const fetchData = useCallback( async (opts) => {
-    try {
-      const response = await fetch(`${url}?${buildTableQueryString(opts)}`);
-      if (response.ok) {
-        const body = await response.json();
-        const {results, ...others} = body;
-        setLoading(false);
-        setData({
-          ...others,
-          data: results,
-        });
-      } else {
-        throw new Error(`Failed to fetch from API: ${response.status} ${response.statusText}`);
-      }
-    } catch (error) {
-      setLoading(false);
-      setError(error);
-    }
-  }, [url, setData, setLoading, setError]);
-
-  return {
-    ...data,
-    loading,
-    error,
-    fetchData,
-  };
-}
+import useFetchData from './useFetchData';
 
 export default function useVariantTranscripts(variantId) {
   return useFetchData(`/api/variant/${variantId}/transcripts`);
